Lots of fixes
- Try to fix resolution of env_projectedtexture
- Added filterSize to env_projectedtexture
- Added BumpWeapon var for weapon swap hud (not working)
- Decrease bloom
- Update Source Shader Editor (buildcubemaps not working anymore)
- Enabled COLLISION_GROUP_INTERACTIVE_DEBRIS to collide with each other
- Ragdoll rollback from modifiying random values
- Removed blood generation on ragdoll creation
- Reduced sharpnel bullets from grenade explosion from 60 to 40
- Fixed weapon slot
- Fixed Armor droping from combine soldier
- Reduced chances of loots from combine soldier & metropolice
- Added ConVar acsmod_autoitempickup
- Enabled prop_breakable & prop_physics to be inited if taking fire
damage
- Fixes name of Bullets Penetration method
- Speed up Lower & Raise time of weapon
- Some stuff from biohazard

+++ sp/src/game/client/detailobjectsystem.cpp
@@ -1478,7 +1478,7 @@ void CDetailObjectSystem::LevelInitPreEntity()
 		PrecacheMaterial( DETAIL_SPRITE_MATERIAL );
 		IMaterial *pMat = m_DetailSpriteMaterial;
 		// adjust for non-square textures (cropped)
-		float flRatio = pMat->GetMappingWidth() / pMat->GetMappingHeight();
+		float flRatio = pMat->GetMappingWidth() / (float)pMat->GetMappingHeight();
 		if ( flRatio > 1.0 )
 		{
 			for( int i = 0; i<m_DetailSpriteDict.Count(); i++ )
